+++ packages/terra-layout/examples/test-examples/LayoutNoHeader.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import Layout from '../../lib/Layout';
+import TestMenu from './test-content/TestMenu';
+import TestContent from './test-content/TestContent';
+
+const LayoutNoHeader = () => (
+  <Layout
+    key="layout-no-header"
+    menu={<TestMenu />}
+    menuText="Menu Text"
+    id="test-root"
+  >
+    <TestContent />
+  </Layout>
+);
+
+export default LayoutNoHeader;