+++ postcss.config.js
@@ -1,3 +1,0 @@
-const config = require('./packages/terra-framework-site/postcss.config');
-
-module.exports = config;