+++ packages/terra-layout/CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased
 ### Added
 * Debounce to resize of layout
 
+###Changed
+* Updating tests to point to new terra-framework-site location.
+
 1.0.0 - (November 17, 2017)
 ------------------
 Initial stable release